updated typedef
