+++ modules/certificate/certificate.graphql.js
@@ -1,7 +1,7 @@
-const certificateType = `
+const certificateTypes = `
   type Certificate {
     _id: ID!
-    createdBy: ID!
+    createdBy: User
     name: String!
     value: Int!
     isUsed: Boolean!
@@ -9,17 +9,30 @@ const certificateType = `
     dateStart: Date
     dateEnd: Date
   }
-`;
+  
+  type PaginatedCertificate {
+      items: [Certificate]
+      count: Int
+  }
 
-const certificateInput = `
   input CertificateInput {
     name: String
     value: Int
     isUsed: Boolean
   }
+  
+  union CertificateResult = Certificate | Error
+
+  extend type Query {
+    getAllCertificates(limit:Int, skip:Int): PaginatedCertificate!
+    getCertificateById(id: ID!): CertificateResult
+  }
+
+  extend type Mutation {
+    addCertificate(certificate: CertificateInput!): CertificateResult
+    deleteCertificate(id: ID!): CertificateResult
+    updateCertificate(name: String!): CertificateResult
+  }
 `;
 
-module.exports = {
-  certificateType,
-  certificateInput,
-};
+module.exports = { certificateTypes };
