+++ typeDefs.js
@@ -70,8 +70,7 @@ const {
   currencyInput,
 } = require('./modules/currency/currency.graphql');
 const {
-  certificateType,
-  certificateInput,
+  certificateTypes,
 } = require('./modules/certificate/certificate.graphql.js');
 const {
   commentType,
@@ -182,7 +181,7 @@ const typeDefs = gql`
 	${categoryType}
 	${paginatedCategory}
   ${currencyType}
-  ${certificateType}
+  
   ${materialType}
   ${newsType}
   ${patternType}
@@ -438,12 +437,7 @@ const typeDefs = gql`
       count: Int
   }
 
-  type PaginatedCertificate {
-      items: [Certificate]
-      count: Int
-  }
-  
-  type PaginatedPockets {
+   type PaginatedPockets {
     items: [Pocket]
     count: Int
   }
@@ -502,7 +496,7 @@ const typeDefs = gql`
   union PaginatedProductsResult = PaginatedProducts | Error
   union PaginatedCommentsResult = PaginatedComments | Error
   union CategoryResult = Category | Error
-  union CertificateResult = Certificate | Error
+  
   union CurrencyResult = Currency | Error
   union MaterialResult = Material | Error
   union PatternResult = Pattern | Error
@@ -627,8 +621,7 @@ const typeDefs = gql`
     ): PaginatedReplies!
     getRecentComments(limit: Int!): [CommentResult]
     getAllCommentsByUser(userId: ID!): [CommentResult]
-    getAllCertificates(limit:Int, skip:Int): PaginatedCertificate!
-    getCertificateById(id: ID!): CertificateResult
+
     getAllBusinessTexts: [BusinessText]
     getBusinessTextById(id: ID!): BusinessTextResult
     getBusinessTextByCode(code: String!): BusinessTextResult
@@ -733,7 +726,7 @@ const typeDefs = gql`
     image: Upload
   }
   ${categoryInput}
-  ${certificateInput}
+  
   ${currencyInput}
   ${materialInput}
   ${newsInput}
@@ -894,10 +887,7 @@ const typeDefs = gql`
     addNews(news: NewsInput!, upload: Upload): NewsResult
     deleteNews(id: ID!): NewsResult
     updateNews(id: ID!, news: NewsInput!, upload: Upload): NewsResult
-    "Certificate Mutation"
-    addCertificate(certificate: CertificateInput!): CertificateResult
-    deleteCertificate(id: ID!): CertificateResult
-    updateCertificate(name: String!): CertificateResult
+    
     "User Mutation"
     registerUser(user: userRegisterInput!, language: Int!): User
 
@@ -1060,6 +1050,7 @@ const typeDefs = gql`
     addProductToWishlist(productId: ID!): WishlistResult
     deleteProductFromWishlist(productId: ID!): WishlistResult
   }
+  ${certificateTypes}
 `;
 
 module.exports = typeDefs;
